Refactor classes into multiple files. Fully implement solve method
